I'm try fix bug let test script flow but bugs so much, so i will test with almost testcase fail

+++ src/main/java/com/springboot/jenka_coffee/repository/ProductRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -63,14 +64,15 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     /**
      * Find products by category and price range with pagination
+     * FIX: Convert Double to BigDecimal for proper comparison
      */
     @Query("SELECT p FROM Product p WHERE " +
             "(:categoryId IS NULL OR p.category.id = :categoryId) AND " +
             "(:minPrice IS NULL OR p.price >= :minPrice) AND " +
             "(:maxPrice IS NULL OR p.price <= :maxPrice)")
     Page<Product> findByCategoryAndPriceRange(@Param("categoryId") String categoryId,
-                                              @Param("minPrice") Double minPrice,
-                                              @Param("maxPrice") Double maxPrice,
+                                              @Param("minPrice") BigDecimal minPrice,
+                                              @Param("maxPrice") BigDecimal maxPrice,
                                               Pageable pageable);
 
     /**
@@ -83,6 +85,7 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     /**
      * Filter products by all criteria with pagination
+     * FIX: Convert Double to BigDecimal for proper comparison
      */
     @Query("SELECT p FROM Product p WHERE " +
             "(:categoryId IS NULL OR p.category.id = :categoryId) AND " +
@@ -91,8 +94,8 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
             "(:keyword IS NULL OR LOWER(p.name) LIKE LOWER(CONCAT('%', :keyword, '%')) OR " +
             "LOWER(p.description) LIKE LOWER(CONCAT('%', :keyword, '%')))")
     Page<Product> findByAllCriteria(@Param("categoryId") String categoryId,
-                                   @Param("minPrice") Double minPrice,
-                                   @Param("maxPrice") Double maxPrice,
+                                   @Param("minPrice") BigDecimal minPrice,
+                                   @Param("maxPrice") BigDecimal maxPrice,
                                    @Param("keyword") String keyword,
                                    Pageable pageable);
 }
